requirements.txt
discord.py==2.6.2
yt-dlp==2025.8.22
spotipy==2.25.1
aiohttp==3.12.15
PyNaCl==1.5.0
python-dotenv==1.0.0